First working version
This initial version can export the whole password store to a keepass
file with fixed name. A subdir of the password store can be chosen
instead by editing the script, and a keyfile added likewise. The master
password for the exported keepass file is prompted in the command line.

Reading the password store is very slow, despite attempting a little bit
of concurrency via async/await mechanisms.